Refactor: separate descriptor from lib.rs
