+++ bdk-ffi/src/lib.rs
@@ -7,24 +7,20 @@ use crate::blockchain::{
 };
 use crate::psbt::PartiallySignedTransaction;
 use crate::wallet::{BumpFeeTxBuilder, TxBuilder, Wallet};
-
+use crate::descriptor::Descriptor;
 use bdk::bitcoin::blockdata::script::Script as BdkScript;
 use bdk::bitcoin::secp256k1::Secp256k1;
-use bdk::bitcoin::util::bip32::{DerivationPath as BdkDerivationPath, Fingerprint};
+use bdk::bitcoin::util::bip32::{DerivationPath as BdkDerivationPath};
 use bdk::bitcoin::{Address as BdkAddress, Network, OutPoint as BdkOutPoint, Txid};
 use bdk::blockchain::Progress as BdkProgress;
 use bdk::database::any::{SledDbConfiguration, SqliteDbConfiguration};
-use bdk::descriptor::{DescriptorXKey, ExtendedDescriptor, IntoWalletDescriptor};
+use bdk::descriptor::{DescriptorXKey};
 use bdk::keys::bip39::{Language, Mnemonic as BdkMnemonic, WordCount};
 use bdk::keys::{
     DerivableKey, DescriptorPublicKey as BdkDescriptorPublicKey,
     DescriptorSecretKey as BdkDescriptorSecretKey, ExtendedKey, GeneratableKey, GeneratedKey,
-    KeyMap,
 };
 use bdk::miniscript::BareCtx;
-use bdk::template::{
-    Bip44, Bip44Public, Bip49, Bip49Public, Bip84, Bip84Public, DescriptorTemplate,
-};
 use bdk::wallet::AddressIndex as BdkAddressIndex;
 use bdk::wallet::AddressInfo as BdkAddressInfo;
 use bdk::{Balance as BdkBalance, BlockTime, Error as BdkError, FeeRate, KeychainKind};
@@ -575,186 +571,6 @@ impl DescriptorPublicKey {
 
     fn as_string(&self) -> String {
         self.descriptor_public_key_mutex.lock().unwrap().to_string()
-    }
-}
-
-#[derive(Debug)]
-struct Descriptor {
-    pub extended_descriptor: ExtendedDescriptor,
-    pub key_map: KeyMap,
-}
-
-impl Descriptor {
-    fn new(descriptor: String, network: Network) -> Result<Self, BdkError> {
-        let secp = Secp256k1::new();
-        let (extended_descriptor, key_map) = descriptor.into_wallet_descriptor(&secp, network)?;
-        Ok(Self {
-            extended_descriptor,
-            key_map,
-        })
-    }
-
-    fn new_bip44(
-        secret_key: Arc<DescriptorSecretKey>,
-        keychain_kind: KeychainKind,
-        network: Network,
-    ) -> Self {
-        let derivable_key = secret_key.descriptor_secret_key_mutex.lock().unwrap();
-
-        match derivable_key.deref() {
-            BdkDescriptorSecretKey::XPrv(descriptor_x_key) => {
-                let derivable_key = descriptor_x_key.xkey;
-                let (extended_descriptor, key_map, _) =
-                    Bip44(derivable_key, keychain_kind).build(network).unwrap();
-                Self {
-                    extended_descriptor,
-                    key_map,
-                }
-            }
-            BdkDescriptorSecretKey::Single(_) => {
-                unreachable!()
-            }
-        }
-    }
-
-    fn new_bip44_public(
-        public_key: Arc<DescriptorPublicKey>,
-        fingerprint: String,
-        keychain_kind: KeychainKind,
-        network: Network,
-    ) -> Self {
-        let fingerprint = Fingerprint::from_str(fingerprint.as_str()).unwrap();
-        let derivable_key = public_key.descriptor_public_key_mutex.lock().unwrap();
-
-        match derivable_key.deref() {
-            BdkDescriptorPublicKey::XPub(descriptor_x_key) => {
-                let derivable_key = descriptor_x_key.xkey;
-                let (extended_descriptor, key_map, _) =
-                    Bip44Public(derivable_key, fingerprint, keychain_kind)
-                        .build(network)
-                        .unwrap();
-
-                Self {
-                    extended_descriptor,
-                    key_map,
-                }
-            }
-            BdkDescriptorPublicKey::Single(_) => {
-                unreachable!()
-            }
-        }
-    }
-
-    fn new_bip49(
-        secret_key: Arc<DescriptorSecretKey>,
-        keychain_kind: KeychainKind,
-        network: Network,
-    ) -> Self {
-        let derivable_key = secret_key.descriptor_secret_key_mutex.lock().unwrap();
-
-        match derivable_key.deref() {
-            BdkDescriptorSecretKey::XPrv(descriptor_x_key) => {
-                let derivable_key = descriptor_x_key.xkey;
-                let (extended_descriptor, key_map, _) =
-                    Bip49(derivable_key, keychain_kind).build(network).unwrap();
-                Self {
-                    extended_descriptor,
-                    key_map,
-                }
-            }
-            BdkDescriptorSecretKey::Single(_) => {
-                unreachable!()
-            }
-        }
-    }
-
-    fn new_bip49_public(
-        public_key: Arc<DescriptorPublicKey>,
-        fingerprint: String,
-        keychain_kind: KeychainKind,
-        network: Network,
-    ) -> Self {
-        let fingerprint = Fingerprint::from_str(fingerprint.as_str()).unwrap();
-        let derivable_key = public_key.descriptor_public_key_mutex.lock().unwrap();
-
-        match derivable_key.deref() {
-            BdkDescriptorPublicKey::XPub(descriptor_x_key) => {
-                let derivable_key = descriptor_x_key.xkey;
-                let (extended_descriptor, key_map, _) =
-                    Bip49Public(derivable_key, fingerprint, keychain_kind)
-                        .build(network)
-                        .unwrap();
-
-                Self {
-                    extended_descriptor,
-                    key_map,
-                }
-            }
-            BdkDescriptorPublicKey::Single(_) => {
-                unreachable!()
-            }
-        }
-    }
-
-    fn new_bip84(
-        secret_key: Arc<DescriptorSecretKey>,
-        keychain_kind: KeychainKind,
-        network: Network,
-    ) -> Self {
-        let derivable_key = secret_key.descriptor_secret_key_mutex.lock().unwrap();
-
-        match derivable_key.deref() {
-            BdkDescriptorSecretKey::XPrv(descriptor_x_key) => {
-                let derivable_key = descriptor_x_key.xkey;
-                let (extended_descriptor, key_map, _) =
-                    Bip84(derivable_key, keychain_kind).build(network).unwrap();
-                Self {
-                    extended_descriptor,
-                    key_map,
-                }
-            }
-            BdkDescriptorSecretKey::Single(_) => {
-                unreachable!()
-            }
-        }
-    }
-
-    fn new_bip84_public(
-        public_key: Arc<DescriptorPublicKey>,
-        fingerprint: String,
-        keychain_kind: KeychainKind,
-        network: Network,
-    ) -> Self {
-        let fingerprint = Fingerprint::from_str(fingerprint.as_str()).unwrap();
-        let derivable_key = public_key.descriptor_public_key_mutex.lock().unwrap();
-
-        match derivable_key.deref() {
-            BdkDescriptorPublicKey::XPub(descriptor_x_key) => {
-                let derivable_key = descriptor_x_key.xkey;
-                let (extended_descriptor, key_map, _) =
-                    Bip84Public(derivable_key, fingerprint, keychain_kind)
-                        .build(network)
-                        .unwrap();
-
-                Self {
-                    extended_descriptor,
-                    key_map,
-                }
-            }
-            BdkDescriptorPublicKey::Single(_) => {
-                unreachable!()
-            }
-        }
-    }
-
-    fn as_string_private(&self) -> String {
-        let descriptor = &self.extended_descriptor;
-        let key_map = &self.key_map;
-        descriptor.to_string_with_secret(key_map)
-    }
-
-    fn as_string(&self) -> String {
-        self.extended_descriptor.to_string()
     }
 }
